Moved counters and scorers to subdirectories based on the entities they operate on (PRs vs Comments) #9
