--- RepoMan/RepoMan/Analysis/Counters/ICounter.cs
@@ -1,7 +1,0 @@
-namespace RepoMan.Analysis.Counters
-{
-    internal interface ICounter
-    {
-        int Count(string s);
-    }
-}
