+++ RepoMan/RepoMan/Analysis/Scoring/Comments/WordCountScorer.cs
