+++ RepoMan/RepoMan/Analysis/Scoring/PullRequests/ApprovalScorer.cs
@@ -1,4 +1,4 @@
-namespace RepoMan.Analysis.Scoring
+namespace RepoMan.Analysis.Scoring.PullRequests
 {
     class ApprovalScorer :
         Scorer
