+++ RepoMan/RepoMan/Analysis/Scoring/PullRequests/BusinessDaysScorer.cs
