+++ RepoMan/RepoMan/Analysis/Scoring/PullRequests/UniqueCommenterScorer.cs
