+++ RepoMan/RepoMan/Repository/RepoWorker.cs
@@ -6,6 +6,7 @@ using Octokit;
 using RepoMan.Analysis;
 using RepoMan.Analysis.ApprovalAnalyzers;
 using RepoMan.Analysis.Counters;
+using RepoMan.Analysis.Counters.Comments;
 using Serilog;
 
 namespace RepoMan.Repository
@@ -17,7 +18,7 @@ namespace RepoMan.Repository
         private readonly string _fullName;
         private readonly IApprovalAnalyzer _approvalAnalyzer;
         private readonly ICommentAnalyzer _commentAnalyzer;
-        private readonly ICounter _wordCounter;
+        private readonly ICommentCounter _wordCounter;
         private readonly IRepositoryHealthAnalyzer _repoHealthAnalyzer;
         private readonly ILogger _logger;
 
@@ -25,7 +26,7 @@ namespace RepoMan.Repository
             IRepoManager repo,
             IApprovalAnalyzer approvalAnalyzer,
             ICommentAnalyzer commentAnalyzer,
-            ICounter wordCounter,
+            ICommentCounter wordCounter,
             IRepositoryHealthAnalyzer repoHealthAnalyzer,
             ILogger logger)
         {
