+++ RepoMan/Scratch/Program.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using RepoMan.Analysis.Counters;
+using RepoMan.Analysis.Counters.Comments;
 using RepoMan.Analysis.Scoring;
 
 namespace Scratch
